将层次化展示疫情图的功能封装成控件
